+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Improvements:
     * compiler: support for variadic function argument passing
     * compiler: NEW operator added
     * compiler: EXTENDS, BASE keywords added (OOP inheritance)
+    * compiler: PROPERTY support added (OOP)
 
 Bug Fixes: